requirements.txt
appnope==0.1.0
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
certifi==2018.4.16
chardet==3.0.4
cloudinary==1.12.0
configparser==3.5.0
decorator==4.3.0
Django==1.8.15
django-contrib-comments==1.8.0
django-crispy-forms==1.7.2
django-extensions==2.0.7
django-fluent-comments==2.0.2
django-registration-redux==1.4
django-resized==0.3.9
django-tag-parser==3.1
django-threadedcomments==1.2
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
idna==2.7
ipython-genutils==0.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
pathlib2==2.3.2
pbr==4.0.3
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pylint==1.9.2
python-akismet==0.4.1
pytz==2018.5
requests==2.19.1
scandir==1.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
wrapt==1.10.11
